Move lightcone radec generation into plugin
